Extract controller + fix unit test
