+++ pkg/controller/remoteca/remoteca_controller.go
@@ -1,0 +1,38 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License;
+// you may not use this file except in compliance with the Elastic License.
+
+package remoteca
+
+import (
+	"github.com/elastic/cloud-on-k8s/pkg/controller/common/operator"
+	"github.com/elastic/cloud-on-k8s/pkg/controller/elasticsearch/remotecluster/remoteca"
+	"github.com/elastic/cloud-on-k8s/pkg/utils/rbac"
+	"sigs.k8s.io/controller-runtime/pkg/controller"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+)
+
+var (
+	name = "remoteca-controller"
+)
+
+// Add creates a new RemoteCa Controller and adds it to the manager with default RBAC.
+func Add(mgr manager.Manager, accessReviewer rbac.AccessReviewer, params operator.Parameters) error {
+	r := remoteca.NewReconciler(mgr, accessReviewer, params)
+	c, err := add(mgr, r)
+	if err != nil {
+		return err
+	}
+	return remoteca.AddWatches(c, r)
+}
+
+// add adds a new Controller to mgr with r as the reconcile.Reconciler
+func add(mgr manager.Manager, r reconcile.Reconciler) (controller.Controller, error) {
+	// Create a new controller
+	c, err := controller.New(name, mgr, controller.Options{Reconciler: r})
+	if err != nil {
+		return c, err
+	}
+	return c, nil
+}
